mapeo de entidades y adicion de dependencia para layout

+++ LP II - Dishly Proyecto Final/dishly/src/main/java/com/dishly/app/model/CarritoItem.java
@@ -10,6 +10,6 @@ public class CarritoItem {
     private int cantidad;
 
     public double getSubtotal() {
-        return plato.getPrecio() * cantidad;
+        return plato.getPrecio().doubleValue() * cantidad;
     }
 }
